+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/bonsai/DebugGethTraceJsonRpcHttpBySpecTest.java
@@ -47,7 +47,8 @@ public class DebugGethTraceJsonRpcHttpBySpecTest extends AbstractJsonRpcHttpBySp
         new String[] {
           "debug-geth/specs/prestate-tracer/diff-mode-true",
           "debug-geth/specs/prestate-tracer/diff-mode-false",
-          "debug-geth/specs/call-tracer"
+          "debug-geth/specs/call-tracer",
+          "debug-geth/specs/4byte-tracer"
         });
   }